+++ src/Config/Value.hs
@@ -7,10 +7,12 @@
 module Config.Value
   ( Section(..)
   , Value(..)
+  , Atom(..)
   ) where
 
 import Data.Text    (Text)
 import Data.Data    (Data, Typeable)
+import Data.String  (IsString(..))
 import GHC.Generics (Generic)
 
 -- | A single section of a 'Value'
@@ -20,11 +22,19 @@ data Section = Section
   }
   deriving (Eq, Read, Show, Typeable, Data, Generic)
 
+-- | Wrapper to distinguish 'Atom' from 'Text' by
+-- type in a configuration.
+newtype Atom = MkAtom { atomName :: Text }
+  deriving (Eq, Ord, Show, Read, Typeable, Data, Generic)
+
+instance IsString Atom where
+  fromString = MkAtom . fromString
+
 -- | Sum type of the values supported by this language.
 data Value
   = Sections [Section]
   | Number   Int Integer -- ^ base number
   | Text     Text
-  | Atom     Text
+  | Atom     Atom
   | List     [Value]
   deriving (Eq, Read, Show, Typeable, Data, Generic)